requirements.txt
Django==5.1.3
djangorestframework     # For building RESTful APIs
django-cors-headers     # To enable CORS (Cross-Origin Resource Sharing)
google-auth             # For Google authentication
google-auth-oauthlib    # For OAuth 2.0 support
google-auth-httplib2    # For HTTP transport for Google Auth
djangorestframework-simplejwt  # For JWT authentication
psycopg2
Pillow
